requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
browserbase==1.4.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.2.0
click==8.3.0
coloredlogs==15.0.1
crewai==0.193.2
crewai-tools==0.76.0
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.115.5
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
google-auth==2.41.1
google-genai==1.45.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.11.3
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonpickle==4.1.1
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lance-namespace==0.0.18
lance-namespace-urllib3-client==0.0.18
lancedb==0.25.2
libclang==18.1.1
litellm==1.74.9
lxml==5.3.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.1.3
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.109.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
Pillow==10.4.0
playwright==1.55.0
portalocker==2.7.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylance==0.38.2
pypdf==6.1.2
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
pytube==15.0.0
pyvis==0.3.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==3.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
stagehand==0.5.4
starlette==0.41.3
sympy==1.14.0
tenacity==9.1.2
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
torch==2.9.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.19.2
typing_extensions==4.15.0
urllib3==2.3.0
uv==0.9.4
uvicorn==0.32.1
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.22.0
youtube-transcript-api==1.2.3
zipp==3.23.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
sqlalchemy
psycopg2-binary
crewai
httpx


pdfminer.six
pillow
lxml
pandas==2.3.3  # Or the latest compatible version you know works
docling==2.57.0
docling-core==2.49.0
docling-parse==4.7.0
docling-ibm-models==3.10.0

# LlamaIndex + PGVector
llama-index>=0.12.0
llama-index-vector-stores-postgres
llama-index-embeddings-huggingface
llama-index-llms-ollama
pgvector
